requirements.txt
multipledispatch==1.0.0
pypng==0.20220715.0
